Abstract new component containing and controlling video filters
